requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
anyio==4.5.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733175639022/work
async-timeout==5.0.1
attrs==25.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
bertopic==0.14.1
bleach==6.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
confection==0.1.5
cycler==0.12.1
cymem==2.0.11
Cython==0.29.37
dask==2023.5.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1722923746907/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.2.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
filelock==3.16.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gensim==4.2.0
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.12.1
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
iniconfig==2.0.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1683289033986/work
ipywidgets==8.1.5
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.7
langcodes==3.3.0
llvmlite==0.41.1
locket==1.0.0
MarkupSafe==2.1.5
matplotlib==3.5.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
narwhals==1.28.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.1
nltk==3.9.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
openai==1.65.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
partd==1.4.1
pathlib_abc==0.1.1
pathy==0.11.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
plotly==6.0.0
pluggy==1.5.0
preshed==3.0.9
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
propcache==0.2.0
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1719274595110/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pydantic==1.10.8
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pynndescent==0.5.13
pyparsing==3.1.4
pytest==8.3.5
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz==2025.1
PyYAML==6.0.2
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1724399083222/work
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==6.4.0
sniffio==1.3.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.8
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
swifter==1.3.4
sympy==1.13.3
tenacity==9.0.0
thinc==8.1.10
threadpoolctl==3.5.0
tokenizers==0.13.2
tomli==2.2.1
toolz==1.0.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1717722826518/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.26.1
triton==3.0.0
typer==0.7.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.2.3
wasabi==0.10.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
webencodings==0.5.1
widgetsnbextension==4.0.13
yarl==1.15.2
zipp==3.20.2
